backend/requirements.txt
# Tech Info Plus - Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
# CORS est intégré dans fastapi
python-dateutil>=2.8.0
bcrypt>=4.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
